separate test snippet to file
